refactor(makeFlat): use internal makeFlat
Addresses PR Comments

Left one as a named function to recurse over
